# grupone/openapi.json
{"openapi":"3.0.1","info":{"title":"Grupo NE Maintenance API","description":"API backend para mantenimiento automotriz con enfoque predictivo","contact":{"name":"Grupo NE"},"license":{"name":"Proprietary"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/vehicles/{id}":{"get":{"tags":["vehicle-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}},"put":{"tags":["vehicle-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}},"delete":{"tags":["vehicle-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1/spare-parts/{id}":{"get":{"tags":["spare-part-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SparePart"}}}}}},"put":{"tags":["spare-part-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SparePart"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SparePart"}}}}}},"delete":{"tags":["spare-part-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1/maintenance-records/{id}":{"get":{"tags":["maintenance-record-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MaintenanceRecord"}}}}}},"put":{"tags":["maintenance-record-controller"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaintenanceRecord"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MaintenanceRecord"}}}}}},"delete":{"tags":["maintenance-record-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1/vehicles":{"get":{"tags":["vehicle-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"post":{"tags":["vehicle-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"/api/v1/spare-parts":{"get":{"tags":["spare-part-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SparePart"}}}}}}},"post":{"tags":["spare-part-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SparePart"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SparePart"}}}}}}},"/api/v1/maintenance-records":{"get":{"tags":["maintenance-record-controller"],"operationId":"getAll_2","parameters":[{"name":"vehicleId","in":"query","required":false,"schema":{"type":"string"}},{"name":"overdue","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MaintenanceRecord"}}}}}}},"post":{"tags":["maintenance-record-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaintenanceRecord"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MaintenanceRecord"}}}}}}},"/api/v1/alerts":{"get":{"tags":["alert-controller"],"operationId":"getAll_3","parameters":[{"name":"openOnly","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Alert"}}}}}}},"post":{"tags":["alert-controller"],"operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Alert"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Alert"}}}}}}},"/api/v1/spare-parts/{id}/stock":{"patch":{"tags":["spare-part-controller"],"operationId":"adjustStock","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"delta","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SparePart"}}}}}}},"/api/v1/alerts/{id}/status":{"patch":{"tags":["alert-controller"],"operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"status","in":"query","required":true,"schema":{"type":"string","enum":["OPEN","ACKNOWLEDGED","RESOLVED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Alert"}}}}}}},"/api/v1/spare-parts/low-stock":{"get":{"tags":["spare-part-controller"],"operationId":"getLowStock","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SparePart"}}}}}}}},"/api/v1/dashboard/summary":{"get":{"tags":["dashboard-controller"],"operationId":"summary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardSummaryResponse"}}}}}}},"/api/v1/alerts/{id}":{"get":{"tags":["alert-controller"],"operationId":"getById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Alert"}}}}}},"delete":{"tags":["alert-controller"],"operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}}}},"components":{"schemas":{"Vehicle":{"required":["brand","model","plate","status","year"],"type":"object","properties":{"id":{"type":"string"},"plate":{"type":"string"},"brand":{"type":"string"},"model":{"type":"string"},"year":{"type":"integer","format":"int32"},"currentMileage":{"type":"integer","format":"int64"},"status":{"type":"string"}}},"SparePart":{"required":["category","minStock","name","sku","stock","unitCost"],"type":"object","properties":{"id":{"type":"string"},"sku":{"type":"string"},"name":{"type":"string"},"category":{"type":"string"},"stock":{"minimum":0,"type":"integer","format":"int32"},"minStock":{"minimum":0,"type":"integer","format":"int32"},"unitCost":{"minimum":0,"type":"number","format":"double"}}},"MaintenanceRecord":{"required":["plannedDate","status","type","vehicleId"],"type":"object","properties":{"id":{"type":"string"},"vehicleId":{"type":"string"},"type":{"type":"string","enum":["PREVENTIVE","PREDICTIVE","CORRECTIVE"]},"plannedDate":{"type":"string","format":"date"},"completedDate":{"type":"string","format":"date"},"mileageAtService":{"type":"integer","format":"int64"},"totalCost":{"type":"number","format":"double"},"sparePartIds":{"type":"array","items":{"type":"string"}},"status":{"type":"string"},"notes":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"Alert":{"required":["message","severity","status","title","vehicleId"],"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"message":{"type":"string"},"vehicleId":{"type":"string"},"maintenanceRecordId":{"type":"string"},"severity":{"type":"string"},"status":{"type":"string","enum":["OPEN","ACKNOWLEDGED","RESOLVED"]},"createdAt":{"type":"string","format":"date-time"},"resolvedAt":{"type":"string","format":"date-time"}}},"DashboardSummaryResponse":{"type":"object","properties":{"totalVehicles":{"type":"integer","format":"int64"},"totalMaintenanceRecords":{"type":"integer","format":"int64"},"plannedMaintenances":{"type":"integer","format":"int64"},"inProgressMaintenances":{"type":"integer","format":"int64"},"completedMaintenances":{"type":"integer","format":"int64"},"lowStockParts":{"type":"integer","format":"int64"},"totalAlerts":{"type":"integer","format":"int64"}}}}}}