# grupone/src/main/java/com/grupone/maintenance/dto/DashboardSummaryResponse.java
package com.grupone.maintenance.dto;

public record DashboardSummaryResponse(
        long totalVehicles,
        long totalMaintenanceRecords,
        long plannedMaintenances,
        long inProgressMaintenances,
        long completedMaintenances,
        long lowStockParts,
        long totalAlerts
) {
}